Organize landing page files
