+++ src/app/(landing)/_sections/footer.tsx
